requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.3
pip~=25.0.1
psycopg2~=2.9.10
python-dotenv~=1.1.0
cryptography~=44.0.2
pillow~=11.1.0
pandas~=2.2.3
pdfplumber~=0.11.6